Safari-611.1.10/LayoutTests/crypto/workers/subtle/ec-generate-key-expected.txt

[Worker] Test generating an EC key pair with P-256 using ECDH algorithm in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-generate-key.js
[Worker] Generating a key pair...
PASS [Worker] keyPair.toString() is '[object Object]'
PASS [Worker] keyPair.publicKey.type is 'public'
PASS [Worker] keyPair.publicKey.extractable is true
PASS [Worker] keyPair.publicKey.algorithm.name is 'ECDH'
PASS [Worker] keyPair.publicKey.algorithm.namedCurve is 'P-256'
PASS [Worker] keyPair.publicKey.usages is [ ]
PASS [Worker] keyPair.privateKey.type is 'private'
PASS [Worker] keyPair.privateKey.extractable is false
PASS [Worker] keyPair.privateKey.algorithm.name is 'ECDH'
PASS [Worker] keyPair.privateKey.algorithm.namedCurve is 'P-256'
PASS [Worker] keyPair.privateKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
